Retarus/pyproject.toml
[tool.poetry]
name = "retarus"
version = "0.1.0"
description = "Retarus intake module"
authors = ["Alexandre Léonardi <alexandre.leonardi@sekoia.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
sekoia-automation-sdk = "^1.3.8"
orjson = "^3.7.7"
websocket-client = "^1.3.3"
python-dateutil = "^2.8.2"
prometheus-client = "^0.16.0"
confluent-kafka = "^2.1.1"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
requests = "*"
requests-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 119

[tool.coverage.run]
omit = [
    "tests/*",
    "main.py",
]
